chore(deviceauth): moving tenantadm client to pkg
Changelog: None
Ticket: MEN-7715
